AndroidKeystore: Improve support for Attest Keys. (#1263)
Introduce Algorithm.ANDROID_KEYSTORE_ATTEST_KEY which can be used to
create an Android Keystore Attest Key which can be used together with
existing AndroidKeystoreCreateKeySettings.setAttestKeyAlias(). This is
an improvement as up to now the application would need to use the
native Android APIs to create an Attest Key.

Also change KeyInfo.attestation for a key using an attest key. The
previous behavior was that the attestation only contained a single
certificate and the new behavior is that the attest key's attestation
chain is appended at the end. This way the attestation won't have to
be special cased and it goes all the way up to the attestation root,
like normal keys.

Also add new items in Android Keystore Secure Area screen in testapp
for this.

Finally, also change error message in CborSymbolProcessor to reference
the class name for schema changes.

Test: Manually tested + unit tests.

Signed-off-by: David Zeuthen <zeuthen@google.com>