NFC: Simplify NFC reading and allow using the tag after the scanning prompt. (#1282)
Currently `scanNfcMdocReader()` returns the result in a lambda with
the semantics that if the created transport was for NFC data transfer,
that lambda is called in a separate coroutine and the caller is
expected to perform all data transfer in this call.

This is a somewhat unintuitive and even undesirable API as an Identity
Reader app will typically want to switch to another screen as soon as
data transfer is initiated and the fact that the launched coroutine is
canceled with switching to another screen this doesn't work in
practice for 18013-5 NFC data transfer.

Thus, change the API of `scanNfcTag()` to return the result in the
return value and make the caller responsible for closing the
`NfcIsoTag`. This change allows for deferred processing even after
`scanNfcTag()` returns. Use this in `scanNfcMdocReader()` and have
`NfcTransportMdocReader()` close the tag when itself is closed. Also
move the function to update text in the Scan NFC dialog to `NfcIsoTag`
itself.

As a result NFC data transfer now works much better both in Multipaz
Test App and also in Identity Reader. This also paves the way for
creating a hybrid 18013-5 transport where we try to stay on NFC as
long as possible and fall back to BLE only if the tag and reader is
separated (see AP 109.04 in ISO JTC1 SC17 WG10).

Also improve "ISO mdoc Proximity Reader" screen in testapp so it jumps
directly to the "Received Credentials" screen in the case of the
connection being closed after a single request. Also show more
detailed "transfer info" including engagement type, transfer protocol,
request/response size, and durations in various stages of the
protocol.

Test: Manually tested on both Android and iOS.

Signed-off-by: David Zeuthen <zeuthen@google.com>