testapp: Use non-backed-up storage. (#1310)
It looks like recent Android versions are backing up and restoring the
app's data directory. This leads to crashes when the app is starting
up and looking for Android Keystore keys that don't exist. This can
easily be reproduced by uninstalling and then reinstalling the
app. For example, if you squash multiple commits into one then -
because we use the number of commits as the version number - you need
to uninstall the old version and reinstall so this is easy to run
into.

Move testapp entirely to storing its data in non-backed-up data to
work around this.

Test: Manually tested.

Signed-off-by: David Zeuthen <zeuthen@google.com>