Mdoc: Use MdocBleServiceData to convey the BLE L2CAP PSM. (#1410)
Previously we were using another mechanism using a not-yet-allocated
number. After conferring with the Bluetooth SIG we converged on
another solution whereby ISO JTC1 SC17 would be assigned a 16-bit
UUID.

The UUID actually hasn't been given to SC17 yet so we're using a
temporary UUID of 0xFF01 for the test event.

Test: Manually tested.

Signed-off-by: David Zeuthen <zeuthen@google.com>