Update to Kotlin 2.2.21 and XCode 26.1.1. (#1428)
Also update all dependencies to their latest version, except for
navigation which will require some extra work.

Also apply a fix to make sure testapp works on the Android emulator
which it hasn't since we added Observe mode support a couple of weeks
back.

This unblocks work on #1364 which is about adding DC API support on
iOS which requires iOS 26.0.

Test: Manually tested on both Android, iOS, and server-side.

Signed-off-by: David Zeuthen <zeuthen@google.com>