Use correct COSE algorithms in 18013-5 DeviceRequest. (#1461)
When we switched to AsymmetricKey we introduced a bug whereby the
wrong COSE Algorithm would be used in the COSE_Sign1 for reader
authentication, for example ESP256 (-9) instead of ES256 (-7). Add
tests for this and fix it. Do this for both `DeviceRequest` and
deprecated `DeviceRequestGenerator`.

Also simplify DeviceRequest.kt codeflows to not call a helper - this
was introduced back before the AsymmetricKey change since we had
multiple user-visible methods calling the same helper. It should
probably have been simplified as part of moving to AsymmetricKey.

Test: Unit tests.

Signed-off-by: David Zeuthen <zeuthen@google.com>